refactor(parametrage): renommer parametres_generaux en configuration_globale
Deux permissions portaient le même titre "Paramètres généraux" dans
l'interface d'admin : portail_parametres_modifier (section Portail) et
parametres_generaux (section Paramétrage), ce qui était ambigu.

- Renommage fichiers: utils_parametres_generaux.py -> utils_configuration_globale.py
- Renommage fichiers: parametrage/forms/parametres_generaux.py -> configuration_globale.py
- Renommage fichiers: parametrage/views/parametres_generaux.py -> configuration_globale.py
- Mise à jour imports dans core/views/base.py, core/views/menu.py, core/utils/utils_permissions.py
- Mise à jour URLs dans parametrage/urls.py
- Mise à jour imports dans fiche_famille et fiche_individu (ajouts de notre branche)
- Renommage variables locales, paramètres de fonctions et cache keys
- Titre du menu et de la page changé en "Configuration globale"
