+++ noethysweb/core/utils/utils_configuration_globale.py
