+++ noethysweb/core/utils/utils_permissions.py
@@ -8,12 +8,12 @@ from fiche_famille.utils.utils_famille import LISTE_ONGLETS as LISTE_ONGLETS_FAM
 from fiche_individu.utils.utils_individu import LISTE_ONGLETS as LISTE_ONGLETS_INDIVIDUS
 
 
-def GetPermissionsPossibles(parametres_generaux=None, organisateur=None, categorie=None):
+def GetPermissionsPossibles(configuration_globale=None, organisateur=None, categorie=None):
     """ Liste des commandes pour créer les permissions des utilisateurs """
     liste_permissions = []
 
     # Commandes de menu
-    menu_principal = GetMenuPrincipal(parametres_generaux=parametres_generaux, 
+    menu_principal = GetMenuPrincipal(configuration_globale=configuration_globale,
                                     organisateur=organisateur, force_permissions=True)
 
     for menu in menu_principal.GetChildren():
