+++ noethysweb/core/views/base.py
@@ -12,7 +12,7 @@ from django.conf import settings
 from django.core.cache import cache
 from core.views.menu import GetMenuPrincipal
 from core.models import Organisateur, Consommation, PortailMessage, PortailRenseignement
-from core.utils import utils_parametres , utils_parametres_generaux
+from core.utils import utils_parametres , utils_configuration_globale
 from noethysweb.version import GetVersion
 from django.db.models import Q
 
@@ -122,11 +122,11 @@ class CustomView(LoginRequiredMixin, UserPassesTestMixin): #, PermissionRequired
         context['organisateur'] = organisateur
 
         # Paramètres du portail
-        parametres_generaux = cache.get('parametres_generaux')
-        if not parametres_generaux:
-            parametres_generaux = utils_parametres_generaux.Get_dict_parametres()
-            cache.set('parametres_generaux', parametres_generaux, 1)
-        context['parametres_generaux'] = parametres_generaux
+        configuration_globale = cache.get('configuration_globale')
+        if not configuration_globale:
+            configuration_globale = utils_configuration_globale.Get_dict_parametres()
+            cache.set('configuration_globale', configuration_globale, 1)
+        context['configuration_globale'] = configuration_globale
 
         # Options d'interface
         key_cache = "options_interface_user%d" % self.request.user.pk
@@ -145,7 +145,7 @@ class CustomView(LoginRequiredMixin, UserPassesTestMixin): #, PermissionRequired
             cache.set(key_cache, parametres)
 
         # Mémorise le menu principal
-        menu_principal = GetMenuPrincipal(parametres_generaux=parametres_generaux, organisateur=organisateur, user=self.request.user)
+        menu_principal = GetMenuPrincipal(configuration_globale=configuration_globale, organisateur=organisateur, user=self.request.user)
         context['menu_principal'] = menu_principal
 
         # Si la page est un crud, on récupère l'url de la liste en tant que menu_code
