+++ noethysweb/fiche_individu/utils/utils_individu.py
@@ -22,7 +22,7 @@ LISTE_ONGLETS = [
 ]
 
 def Get_filtered_onglets():
-    from core.utils.utils_parametres_generaux import Get_dict_parametres
+    from core.utils.utils_configuration_globale import Get_dict_parametres
     parametres = Get_dict_parametres()
 
     # Filtrer la liste en fonction des paramètres
