+++ noethysweb/parametrage/urls.py
@@ -27,7 +27,7 @@ from parametrage.views import organisateur, structures, \
     types_qualifications_collaborateurs, types_pieces_collaborateurs, types_evenements_collaborateurs, types_postes_collaborateurs, \
     modeles_plannings_collaborateurs, groupes_collaborateurs, modeles_aides, transports, compagnies, lignes, lieux, arrets, modeles_impressions, \
     modeles_word, releves_bancaires, sondages, achats_categories, achats_fournisseurs, modeles_commandes, modeles_commandes_colonnes, \
-    activites_evenements_categories, activites_import_export, api_particulier, parametres_generaux
+    activites_evenements_categories, activites_import_export, api_particulier, configuration_globale
 
 
 urlpatterns = [
@@ -40,7 +40,7 @@ urlpatterns = [
     path('parametrage/organisateur/modifier/<int:pk>', organisateur.Modifier.as_view(), name='organisateur_modifier'),
 
     # Paramètres généraux
-    path('parametrage/parametres_generaux/', parametres_generaux.Modifier.as_view(), name='parametres_generaux'),
+    path('parametrage/configuration_globale/', configuration_globale.Modifier.as_view(), name='configuration_globale'),
     # Structures
     path('parametrage/structures/liste', structures.Liste.as_view(), name='structures_liste'),
     path('parametrage/structures/ajouter', structures.Ajouter.as_view(), name='structures_ajouter'),
